--- cmd/ocagent/config.yaml
@@ -1,17 +1,0 @@
-receivers:
-    opencensus:
-        address: "127.0.0.1:55678"
-
-    zipkin:
-        address: "127.0.0.1:9411"
-
-exporters:
-    stackdriver:
-        project: "project-id" # Optional if on GCP, defaults to agent project
-        enable_tracing: true
-
-    zipkin:
-        endpoint: "http://127.0.0.1:9411/api/v2/spans"
-
-zpages:
-    port: 55679
