+++ cmd/occollector/app/collector/testdata/otelsvc-config.yaml
