+++ demos/trace/docker-compose.yaml
@@ -15,11 +15,11 @@ services:
       - "9411:9411"
 
   # Collector
-  oc-collector:
-    image: occollector:latest
-    command: ["--config=/etc/oc-collector-config.yaml", "--http-pprof-port=1777"]
+  otelsvc-collector:
+    image: otelsvc:latest
+    command: ["--config=/etc/otelsvc-collector-config.yaml", "--http-pprof-port=1777"]
     volumes:
-      - ./oc-collector-config.yaml:/etc/oc-collector-config.yaml
+      - ./otelsvc-collector-config.yaml:/etc/otelsvc-collector-config.yaml
     ports:
       - "55678"
       - "55680:55679"
@@ -30,24 +30,24 @@ services:
       - zipkin-all-in-one
 
   # Agent
-  oc-agent:
-    image: ocagent:latest
-    command: ["--config=/etc/oc-agent-config.yaml", "--http-pprof-port=1888"]
+  otelsvc-agent:
+    image: otelsvc:latest
+    command: ["--config=/etc/otelsvc-agent-config.yaml", "--http-pprof-port=1888"]
     volumes:
-      - ./oc-agent-config.yaml:/etc/oc-agent-config.yaml
+      - ./otelsvc-agent-config.yaml:/etc/otelsvc-agent-config.yaml
     ports:
       - "1888:1888"
       - "14268"
       - "55678"
       - "55679:55679"
     depends_on:
-      - oc-collector
+      - otelsvc-collector
 
   # Synthetic load generator
   synthetic-load-generator:
     image: omnition/synthetic-load-generator:1.0.25
     environment:
-      - JAEGER_COLLECTOR_URL=http://oc-agent:14268
+      - JAEGER_COLLECTOR_URL=http://otelsvc-agent:14268
     depends_on:
-      - oc-agent
+      - otelsvc-agent
